+++ R/0imports.R
@@ -7,5 +7,6 @@
 #' @importFrom rlang is_scalar_logical is_scalar_character is_integerish
 #' @importFrom utils write.table tail RShowDoc
 #' @import shinythemes
+#' @import data.table
 options(useFancyQuotes = FALSE)
 NULL